requirements.txt
alembic==1.4.1
aniso8601==8.0.0
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.1
